refactor configuration setup for more sensical require loading
