+++ lib/farscape/helpers.rb
@@ -1,7 +1,9 @@
+require 'farscape/configuration'
+
 module Farscape
   module Helpers
     def config
-      Farscape.config
+      Farscape::Configuration.config
     end
   end
 end
